requirements.txt
alembic==1.16.4
APScheduler==3.11.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
Flask==3.1.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyMySQL==1.1.1
pytest==8.4.1
python-dotenv==1.1.1
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
Werkzeug==3.1.3
